code/urbanflood_larfno/pyproject.toml
[project]
name = "neuraloperator"
description = "NeuralOperator: Learning in Infinite Dimensions"
authors = [
    {name = "Jean Kossaifi <jean.kossaifi@gmail.com>, David Pitt <david.h.pitt@gmail.com>, Nik Kovachki, Zongyi Li", email="david.h.pitt@gmail.com"},
]
dependencies = [
    "ruamel-yaml",
    "configmypy",
    "tensorly",
    "tensorly-torch",
    "torch-harmonics==0.7.3",
    "matplotlib",
    "numpy>=1.25",
    "opt-einsum",
    "h5py",
    "zarr",
]
requires-python = ">=3.9"
readme = "README.rst"
license = {text = "MIT"}

# Dynamic versioning
dynamic = ["version"]

# Required config for build system
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["neuralop"]

# Dynamic versioning from neuralop.__init__
[tool.setuptools.dynamic]
version = {attr = "neuralop.__version__"}

[project.optional-dependencies]
dev = [
    "pytest",
    "flaky",
    "black"
]
doc = [
    "matplotlib",
    "myst-nb",
    "numpydoc",
    "numpy>=1.25",
    "sphinx",
    "sphinx-gallery",
    "tensorly_sphinx_theme",
    "torchtnt",
]
all = [
    "pytest",
    "flaky",
    "black",
    "matplotlib",
    "myst-nb",
    "numpydoc",
    "sphinx",
    "sphinx-gallery",
    "tensorly_sphinx_theme",
    "torchtnt",
    "numpy>=1.25",
    "scipy",
]
